feat add loan a book feature

+++ frontend/src/app/members/page.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import { ListTable } from "@/components/ListTable";
+import MyAlertDialog from "@/components/MyAlertDialog";
 import { Button } from "@/components/ui/button";
 import { useToast } from "@/hooks/use-toast";
 import { deleteMember, getMembers, IMember } from "@/services/memberService";
@@ -13,6 +14,7 @@ const Members = () => {
   const { toast } = useToast();
 
   const [members, setMembers] = useState<IMember[]>([]);
+  const [isOpen, setIsOpen] = useState(false);
 
   const fetchMembers = async () => {
     const response = await getMembers();
@@ -27,16 +29,22 @@ const Members = () => {
     router.push(`/members/manage/${memberId}`);
   };
 
+  const showAlertDialog = () => {
+    setIsOpen(true);
+  };
+
   const handleDelete = async (memberId: string) => {
     try {
       await deleteMember(memberId);
 
+      setIsOpen(false);
       toast({
         title: "Member deleted successfully",
         description: "Member has been deleted successfully.",
       });
       fetchMembers();
     } catch {
+      setIsOpen(false);
       toast({
         title: "Failed to delete member",
         description: "An error occurred while deleting member.",
@@ -58,15 +66,20 @@ const Members = () => {
           <Pencil />
         </Button>
         <span className="mx-2"></span>
-        <Button
-          variant={"outline"}
-          onClick={() => handleDelete(member._id as string)}
-        >
+        <Button variant={"outline"} onClick={showAlertDialog}>
           <Trash2 />
         </Button>
+        <MyAlertDialog
+          isOpen={isOpen}
+          title="Delete Member"
+          description="Are you sure you want to delete this member?"
+          onCancel={() => setIsOpen(false)}
+          onAction={() => handleDelete(member._id as string)}
+        />
       </>,
     ],
-    onRowClick: () => handleOnRowClick(member._id as string),
+    onRowClick: () => {},
+    //handleOnRowClick(member._id as string)
   }));
 
   return (
@@ -74,14 +87,15 @@ const Members = () => {
       <Button
         variant={"outline"}
         onClick={() => router.push("/members/manage")}
-        className="m-4">
+        className="m-4"
+      >
         <UserRoundPlus />
       </Button>
       <ListTable
         caption="Members"
         headers={["No", "Name", "Email", "Borrowed Books", "Actions"]}
         rows={rows}
-        tooltip="Click to see this member's detail"
+        tooltip="Click edit to see this member's detail"
       />
     </div>
   );
